requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile ../requirements.in
#
cffi==1.15.1
    # via cryptography
cryptography==41.0.1
    # via -r ../requirements.in
pycparser==2.21
    # via cffi
sqlcipher3-binary==0.5.0
    # via -r ../requirements.in
